refactor to Privacy Policie and Terms and Conditions
